Re-write entire structure
A clean up of 'cstr' design

old 'cstr' design is too vulnerable to bug and overflow. There is no
effective mechanism to decide the type of a new string and avoid
overwrite something

As a result, old 'cstr' design is reject for new struct design. This
allow flexible type decision on multiple machine and settings