Add wchar port
Add wide character support

Wide character type (wcstr_t) and its library are mostly (if not,
completely) the same as single character equivalent. Even the inner
struct they use are the same
The only major change is that there is 3 flag: __need_struct,
__need_cstr_inner_func and __need_wcstr_inner_func.
If define __need_struct, it will reveal inner structure. If define both
__need_struct and __need_cstr_inner_func (or __need_wcstr_inner_func),
it will reveal all of inner function. This is for usage in library
source file and should not be exposed.