Optiomize __cstr_nof_buffer
Replace division operator with right shift operator
__cstr_nof_buffer() provide the number of *buffer*, which is a large
amount of memory to be allocate and access. *inline* __cstr_nof_buffer() is call on every operation that involve allocation. It do this by calculate the size (capacity) to allocate divided by the buffer size of certain type. The division, even optiomized, are very slow(on the scale of CPU clocks)
This commit replace the need of division operator with right shift
operator. This add Tx_MASK, where x is type number, be it 0, 1 or 2.
Since the buffer size is power of 2, this reduce significant calculate
time.